✨ feat: 작품 감상하러 가기 버튼 연결
