+++ App.tsx
@@ -2,7 +2,6 @@ import {createNativeStackNavigator} from '@react-navigation/native-stack';
 import { NavigationContainer } from '@react-navigation/native';
 import Login from './src/screens/User/Login'
 import Signup from './src/screens/User/Signup'
-import MainScreen from './src/screens/Main/MainScreen';
 import DescriptionScreen from './src/screens/Main/DescriptionScreen';
 import FocusPointingScreen from './src/screens/Main/FocusPointingScreen';
 import Image2VideoScreen from './src/screens/Main/Image2VideoScreen';
@@ -24,7 +23,6 @@ const App = () => {
            */}
 
           <Stack.Screen name="HomeScreen" component={HomeScreen} />
-          <Stack.Screen name="MainScreen" component={MainScreen} />
           <Stack.Screen name="ScanImage" component={ScanImage} /> 
           <Stack.Screen name='NDescriptionScreen' component={NDescriptionScreen} />
           <Stack.Screen name="FocusPointingScreen" component={FocusPointingScreen} />
